fleet/requirements.txt
blinker==1.9.0
flask-cors==5.0.1
click==8.3.3
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.12
python-dotenv==1.2.2
Werkzeug==3.1.8
